new commands to separate file, debugging
